Fix bugs

+++ mgcn/utils.py
@@ -82,7 +82,7 @@ def load_features_labels(layer_num, path, dataset,N=-1):
     labels = encode_onehot(idx_features_labels[:, -1])
 
     #build graph
-    idx = np.array(idx_features_labels[:, 0], dtype=np.float64).astype(np.int32)
+    idx = np.array(idx_features_labels[:, 0], dtype=np.float32).astype(np.int32)
     idx_map = {j: i for i, j in enumerate(idx)}
 
     features = normalize(features)
